feat(editor): implemented alignment tools (#1099)
Implemented alignment tools for manual layouting. Alignment applied to
leaf nodes only.

![alignment-tools](https://github.com/user-attachments/assets/9d6e5b1b-e2af-4cce-bac5-b0a3a1dfc73b)


Also replaced @lume/kiwi solver with simple implementation. kiwi solver
does not work properly for compound nodes on 2+ level with a single
child. I guess that was the reason why dragging was disabled for nodes
which are the only child. This workaround is acceptable for drag
scenario but does not fit to align scenario: it is still required to
align a node with others, even it is the only child.
Although implementation with solver is more flexible than bounding box
implementation, this flexibility is not used.