feat: global element and relation predicates
This feature is intended for reusable include / exclude sequences

Architecture documentation can contain multiple views sharing the same
set of repeated include and exclude sequences. Instead of repeating
this set, the maintenance gets easier when the set is defined once and
applied to all relevant views.

This feature allows for definig "global predicate groups" that can
contain a list of include and exclude predicates. Each group can be
explicitly applied to any view.