fix: boundaries of a parent node calculated incorrectly on manual layout (#1520)
The issue appeared after update xyflow/react. Works fine up to (including) the version 12.4.1.
In the version 12.4.2 the update mechanism for InternalNodes was refactored in order to make InternalNode immutable (xyflow/xyflow@7cb9a86).

Workaround is based on assumption that value of node.internals.positionAbsolute the field is incorrect while node.position is correct. Therefore, absolute position of the node's rectangle is calculated from the position of its parent and node's relative coords.

Closes #1516