chore: release v1.23.1 (#1636)
This pull request prepares the following release:
| Repository | Branch | Update | Change |
|---|---|---|---|
| [likec4](https://github.com/likec4/likec4) | releases/v1.23.1 | patch
| [`1.23.0` →
`1.23.1`](https://github.com/likec4/likec4/compare/v1.23.0...releases/v1.23.1)
|

### Release Summary
This is going to be published with the following command:

| Dir | Command |
|---|---|
| packages/core | pnpm publish --tag latest --access public |
| packages/log | pnpm publish --tag latest --access public |
| packages/icons | pnpm publish --tag latest --access public |
| packages/diagram | pnpm publish --tag latest --access public |
| packages/language-server | pnpm publish --tag latest --access public |
| packages/layouts | pnpm publish --tag latest --access public |
| packages/likec4 | pnpm publish --tag latest --access public |

### Merging Instructions
When merging this pull request, you need to **Squash and merge** and
make sure that the title starts with `chore: release v1.23.1`.
<details>
<summary>See details</summary>

After that, a commit `chore: release v1.23.1` will be added and you or
your CI can run `shipjs trigger` to trigger the release based on the
commit.
![Squash and
merge](https://raw.githubusercontent.com/algolia/shipjs/v0.27.0/assets/squash-and-merge.png)
</details>

---

_This pull request is automatically generated by
[Ship.js](https://github.com/algolia/shipjs)_.

---------

Signed-off-by: Denis Davydkov <denis@davydkov.com>